Added 32bit fixes to review

+++ Source/API/lib3mf_reader.cpp
@@ -200,7 +200,7 @@ void Lib3MF::Impl::CReader::AddKeyWrappingCallback(const std::string &sConsumerI
 		if (result == 0)
 			throw ELib3MFInterfaceException(LIB3MF_ERROR_CALCULATIONABORTED);
 
-		plain.resize(needed, 0);
+		plain.resize((size_t)needed, 0);
 
 		result = 0;
 		(*pTheCallback)(entityHandle, cipher.size(), cipher.data(),
